Fix reporting issues  - Disable delete capabilities  - Alter the processing of editing product details  - Add available field to track the availability of product  - Create a generic method to load sales event data together with    related customer and product record.  - Set finished product's availability value to false  - Create setRelatedCustomers and setRelatedProducts methods to    initiallize relatedCustomers and relatedProducts in the salesEventAdapter  - Prevent user from performing executing either the login or signup    operation while the system is trying to login or sigup the user to    the system.

+++ app/src/main/java/com/daveace/salesdiary/fragment/InventoryFragment.java
@@ -121,7 +121,7 @@ public class InventoryFragment extends BaseFragment {
         }
         setLoading(true);
         String name = productNameInput.getText().toString().trim();
-        Integer quantity = Integer.parseInt(quantityInput.getText().toString().trim());
+        double quantity = Double.parseDouble(quantityInput.getText().toString().trim());
         double cost = Double.parseDouble(costInput.getText().toString().trim());
         String code = productCode.getText().toString().trim();
         String imgPath = !(TextUtils.isEmpty(imageFilePath)) ? imageFilePath : "";
